feat: TET-894 unify components folders
