--- src/components/CheckboxGroup/stylesBuilder/index.ts
@@ -1,1 +1,0 @@
-export { stylesBuilder } from './stylesBuilder';
